sbf/sbf/backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Pydantic
pydantic>=2.5.0
pydantic-settings>=2.1.0

# LangGraph and LangChain
langgraph>=0.0.40
langchain>=0.1.0
langchain-core>=0.1.0

# OpenAI
openai>=1.12.0

# Scrapfly
scrapfly-sdk>=0.8.0

# HTML Processing
beautifulsoup4>=4.12.0
lxml>=5.0.0

# PDF Processing
pypdf>=3.17.0
pdfminer.six>=20221105

# Vector Store
chromadb>=0.4.22

# Token counting
tiktoken>=0.5.0

# Rate limiting
slowapi>=0.1.9

# Logging
structlog>=24.1.0

# Async utilities
httpx>=0.26.0
aiofiles>=23.2.1

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Type checking
mypy>=1.8.0

# Linting
ruff>=0.1.0
